refactor fetchCartGrandTotal
