+++ assets/js/theme/custom-add-to-cart.js
@@ -2,6 +2,7 @@ import 'regenerator-runtime';
 import utils from '@bigcommerce/stencil-utils';
 import { getSingleProductQuery } from './get-single-product-query';
 import { getProductsQuery } from './get-products-query';
+import { fetchCartGrandTotal } from './custom-cart-nav-grand-total';
 
 const fetchGraphQLData = async query => {
     const queryOptions = {
@@ -27,25 +28,6 @@ const fetchGraphQLData = async query => {
         throw error;
     }
 };
-
-const fetchCartGrandTotal = () => {
-    var xhr = new XMLHttpRequest();
-    xhr.open('GET', '/cart.php', true);
-    xhr.onreadystatechange = function() {
-        if (xhr.readyState === 4) {
-            if (xhr.status === 200) {
-                var parser = new DOMParser();
-                var doc = parser.parseFromString(xhr.responseText, 'text/html');
-                var grandTotalElement = doc.querySelector('.cart-total-grandTotal span');
-                var grand_total = grandTotalElement ? grandTotalElement.textContent : "$0.00";
-                document.getElementById('custom-total-under-cart').innerHTML = grand_total;
-            } else {
-                document.getElementById('custom-total-under-cart').innerHTML = '$0.00';
-            }
-        }
-    };
-    xhr.send();
-}
 
 const getProducts = async (productIDs) => {
     const query = getProductsQuery(productIDs);
